Use BINDIR, MANDIR and SYSCONFDIR vars in Makefile
This commit introduces these variables, and defaults to using
the same values as the ones that were used before. The advantage
of doing this is to make it easier to use some distro-specific
layout. When this is desirable, these variables can be overridden
to what is appropriate at package build time.