Don't quote ${EDITOR:-vi}
Attached is a trivial patch that removes the quotes from the line that
invokes $EDITOR. It's perfectly cromulent to set $EDITOR to something
with spaces in it, so when we evaluate this one we want it evaluated
bare. For example security nerds might want EDITOR='vim -n' if they
are scared of swap files, and that breaks if we quote there.