Use getopt to regularize options
I like being able to specify command line options (like -c) at the end
of the line (which is usually when I think of the fact that I need
them).

The attached patch uses getopt(1) to regularize the option list so that
lazy people like me can specify the options in any order.