use gpg --batch for quieter output
Right now, every time I call pass to decrypt a key, I get output like:

  You need a passphrase to unlock the secret key for
  user: "User Name <user@domain.com>"
  2048-bit ELG-E key, ID XXXXXXNX, created 2012-04-20 (main key ID NNXXNNNX)
  password

This patch cleans it up so that only the password is output. I use
pinentry-gtk-2 to enter the gpg passphrase, so this information is
redundant to me. I haven't tried other pinentry versions to see if they
repeat the information as well.