Importer for 1Password
An importer script for 1Password. It supports 1Password's text exports
(CSV or TSV) and its 1PIF file format (pseudo-JSON). In addition to the passwords
it imports notes, as well as the username and URL which it stores in passff-
compatible format (it can also use either the title or the URL itself as pass-name).