Allow options in EDITOR variable
This is required by most graphical editors, as the shell command they
provide needs to be told to wait until the file has been closed.

This reverts '311c98c9140205f32be2720db9261c67da658c4c'. We thus use
PATH so that the tests still pass in directories with spaces.