clip: Show an error message if xclip returns a non-zero exit code
xclip will return a non-zero exit code if $DISPLAY is not set, which
might happen, for example, if you're running pass in a tmux session
remotely or something like that. xclip itself will also show an error
message ("Can't open display") which provides enough context to figure
out what the problem is.

There might be other reasons for xclip to return a non-zero exit code,
but the man page doesn't list them.