emacs: Drop dependency on f library
The "f" library is a rather thin translation layer for already
existing Emacs functions.  Most functions directly map to an already
existing function (eg. "f-no-ext" and "file-name-sans-extension").  For
this reason, removing "f" comes at no cost while reducing the number
of dependencies one has to count on and the user has to install.

Co-authored-by: Tino Calancha <tino.calancha@gmail.com>