fish-completion: force some variables to be script-local
Unfortunately, a command "set x" without explicit scope overwrites the variable
"x" in the innermost scope it is defined in, if any.  This can cause problems
if the user defines the variable "x" as global or universal variable (which is
visible in all fishes).  Make sure to define a local variable so we use that.