Add reencrypt command based on reencrypt_path()
`passage reencrypt` let's you reencrypt .age files to add recipients.
Although the same mechanism will also remove access, this is ineffective
when used in combination with git, because the one can just retrieve the
previous version from history.

This functionality corresponds to the second use of `pass init`. Given
that the first (init) functionality is not implemented and internally
this command is not much more than direct access to the
`reencrypt_path() function, `reencrypt` seemed to be a more suitable
name.

The current implementation does not figure out the change in recipients
(if any) and therefore has a rather generic "Reencrypted $path" git
commit message.